Renames formContainer for consistency and moves LeafletMap to a component
