+++ src/components/App.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
-import LeafletMap from '../containers/LeafletMap';
-import FormContainer from '../containers/FormContainer';
+import LeafletMap from './LeafletMap';
+import FormContainer from '../containers/Form';
 
 class App extends Component {
   componentDidMount() {
